requirements.txt
anyio==4.4.0
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.8.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.8.0rc2
distlib==0.3.8
distro==1.7.0
Django==5.0.7
django-allauth==0.63.4
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
environ==1.0
exceptiongroup==1.2.1
fasteners==0.14.1
filelock==3.14.0
Flask==3.0.3
future==0.18.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==2.10
importlib-metadata==4.6.4
inflection==0.5.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.4.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.1
paramiko==2.9.3
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.1
protobuf==3.12.4
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyTelegramBotAPI==4.20.0
python-dateutil==2.8.1
python-decouple==3.8
python-telegram-bot==21.3
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.26.1
wadllib==1.3.6
Werkzeug==3.0.3
xdg==5.0.0
zipp==1.0.0
